+++ tests/test_main.py
@@ -4,6 +4,7 @@ import re
 from datetime import datetime
 from uuid import uuid4
 
+import pendulum
 import pytest
 from beartype import beartype
 from cerberus import Validator
@@ -91,7 +92,7 @@ def test_assert_against_cache_failure(fix_tmp_assert):  # noqa: AAA01
             ],
         ),  # Check UUID types
         (
-            {'dates': str(datetime.now())}, {'dates': str(datetime.utcnow())}, [
+            {'dates': str(datetime.now())}, {'dates': str(pendulum.now().add(weeks=1))}, [
                 KeyRule(pattern=['dates'], func=check_type),
             ],
         ),  # Check date types