+++ tests/test_plugin.py
@@ -23,6 +23,7 @@ def test_assert_against_cache_failure(fix_test_cache, assert_against_cache):
         assert_against_cache({'result': True}, **kwargs)
 
 
+# FIXME: Convert to hypothesis
 @pytest.mark.parametrize(
     ('value', 'expected'), [
         ('20211101', '20211101'),